Fix for bundle issues.
It used to be that every window that had a display in it was a main window, i.e., IdvWindow.getIsAMainWindow returns true, and windows such as the Dashboard were not.  This was changed so only the main window with respect to the TabbedUIManager, the one with the file menu, returned true.  The bundle persistence only persisted windows where getIsAMainWindow() == true.  I overrode IdvUIManager.getWindowsToPersist() to return a list of the windows associated with each display.

Another issue came up where when a bundle was loaded the main window would disappear.  It was just hidden and therefore could be show by using the CRTL-0 key combo, but still was an issue.  As a temporary hack until I can track this issue down I simply overrode IdvUIManager.unpersistWindowInfo() to re-show the window after persistence.