added setSizeBackground.  Works for both buildWindow and openBundle in the background without popping up windows.  Major outstanding issue is that it doesn't handle layer labels at all, which turns out to me more complicated than it should be.. still thinking about how to deal with this.
Also I added a hook in setSize to call setSizeBackground when in offscreen mode.

I have not added height/width to captureImage *yet*... but setSizeBackground should facilitate that in the future if I can get it fully working. So right now, to test setSizeBackground, you need to explicitly call setSize from a background script.  [845]