First attempt at dealing with issue of imagery not 'filling' the display: I added a boolean 'fillDisplay' parameter to createLayer.  Example:
layer1 = panels[0].createLayer('Image Sequence Display', data, fillDisplay=True)

The idea is that when fillDisplay=True, the display will be resized to the size
of the image being displayed, and the projection will be set so that the image
'fills' the entire display with no borders.

However, there are currently a few issues:
- From Jython Shell, there are problems with small images (maybe ~400x400 or smaller).
This seems to be related to setSize not wanting to go so small that UI buttons
will be hidden.
- From Jython Shell, there are problems if image is larger than your screen.
- Only works for 'native' satellite projection (but maybe this is the only
case where we need this functionality? For non-native projections, the image
itself is no longer rectangular...)
- Small (1-2 pixel) white or black borders when you do a captureImage

[845]