Support for executable mcidasv.jar. [1423]
Copies *all* of the JARs McIDAS-V depends upon into "dist/", builds
local-idv.jar and mcidasv.jar, and includes *all* JARs in the
"Class-Path" of mcidasv.jar's MANIFEST.MF.

This means that if you run "java -Xmx1024m -jar mcidasv.jar" or
double-click on mcidasv.jar, Java will search the current directory for
all the JARs specified by "Class-Path"–put another way, no need to
specify classpaths to run McV. This simplifies day-to-day stuff for
developers, but also makes shipping "McIDAS-V.app" and "mcidasv.exe"
that much easier.

There are a couple drawbacks, however. First is that you end up with
copies of all McV jars in your "dist" directory--though this is smart
enough to only copy files when necessary, and they go away if you run
"ant clean". Second is that JAR indexing has to be disabled for this to
work, so there may be a slight perf hit.