Added decorator that forces code to run in EDT [1303]
Jython methods/functions can be executed within the EDT by adding the
"@swing_thread_required" decorator just before the method/function is
defined. Suggestions for a better name are very welcome!

The code currently checks to see if the current thread is the EDT; if
so the decorated code is run immediately (which may be a little too
opportunistic?), otherwise the code gets queued up for execution via
SwingUtilities.invokeLater().