Add "temporary" parameter to setLogLevel. [1303]
Users can now control whether or not setLogLevel will save a given
level between sessions. The default behavior is to consider all
setLogLevel changes as temporary. Saving the log level change can be
accomplished with something like:

  setLogLevel('TRACE', temporary=False)

One thing to be aware of is that saving the log level can only be used
with the "ROOT" logger. So something like the following will raise a
ValueError exception:

  setLogLevel('TRACE', loggerName='ucar.grib2', temporary=False)