Fix for pasting Jython Shell text into "rich" editors.
Bob and Rick (and myself) were noticing that sometimes copying text
from the Jython Shell and pasting into certain editors would munge
newlines. The apparently random behavior is even worse than Amateur
Hour stuff. :(

The problem (in a general sense) turned out to be the fact that I was
"sending" data to the OS clipboard and setting the support content
types to plain text *and* HTML. So-called "rich" text editors (Word,
Wordpad, Thunderbird, Mail.app, etc) were sensibly grabbing the HTML.

The fix is to only send plain text to the clipboard (but this may
become an annoyance if syntax highlighting  ever gets off the ground).

[1303]