Override clone in _MappedAreaImageFlatField. [1303]
The "self * 1" hack is apparently required due to an issue with Jython
proxy objects not working well with Java's "clone". Relying on the
clone from CachedFlatField will actually clone the underlying Java
object(s) as expected…but Jython's __proxy field will point to the
first Jython proxy in both the original Java object as well as the
cloned object!