Change paths to logback config files. [1628]
I'm not wild about the new location of the XML files, but it's a start.
You'll probably want to run "ant clean" prior to building.

As of this commit, McV will look for a config file in two places, in
the following order:

1. build.xml's "logback.default.config" variable. Currently defaults to
"/Users/${user.name}/Documents/McIDAS-V". I'll try to work out
something better for Windows and Linux folks. This is the file that
you'd customize to suit your needs.

2. mcidasv.jar/logback.xml. This is the file that specifies the default
logging behavior.