Simply classpath in startup scripts. [1645]
build.xml has been including the McV classpath within mcidasv.jar's
manifest file for a few months now. Things seem to work.

The nice thing about this approach is that all one has to do to launch
McV is to make sure all of the JARs are in the same directory and then
run "java -jar mcidasv.jar" and things Just Work (though you'd want to
specify a nicely-sized heap!).