Add JVM flag field to startup prefs. [1534]
This is still a work in progress!

This adds the “JVM_OPTIONS” setting to runMcV.prefs, which allows the
user to supply arbitrary things to the JVM…though I’m leaving this
ability out of the startup scripts for now. One thing we’re going to
need to handle/document is which flags the user should not use (things
like -Xmx=1337m).

Also did some cleanup and javadocs.