Add DataTree search to formula param dialog.
This takes the DataTree searching functionality from DataSelector and
generalizes it into the “DataTreeSearcher” class. Each DataTree in the
formula dialog currently has its own DataTreeSearcher.

The basic searching functionality is pretty much working as
expected…provided you limit your interaction with the search/cancel
buttons to mouse clicks. There’s *apparently* an issue with the
keyboard focus that causes the “DataTreeSearchers” to stop responding
to the “control + f” or “cmd + f” hotkeys. :(

Also…apologies for not being able to get this out before the dev
meeting ended! :disappointed:

[1542]