Fix confusion between place/location keywords.
This was due to me misinterpreting what Bob had pretty much spelled
out. Sorry ‘bout my brain. :(

If the location value isn’t a tuple with two elements, the location
will default to (0, 0)…I opted for this approach because the location
is merely optional, but changing the behavior to throw a ValueError or
something would be trivial.

[1665]