Following through on my "bundle version" checker. [644]
This adds the appropriate logic for notifying the user that they are
attempting to load a bundle created with a more recent McIDAS-V—but
only if McV actually generated exceptions while loading the bundle.

The actual notifications are limited to logging warning messages for
now. I didn’t really like showing a simple dialog box with an “UPGRADE
MCV PLZ” message, but maybe it would suffice?

Finally! Here’s the commit ID where the versions get added to bundles:

7291feef34a4a2511c647e838dd2d182eaa8a1c5