Fix copying empty mcidasv.log with fresh user dir.
This commit adds logic to check for a zero-length “mcidasv.log” (prior
to copying into archived_logs). If the file is zero-length, it is
neither copied nor truncated. If the file is not zero-length, the old
copy+truncate stuff happens.

FWIW I’ve only been able to replicate the problem on Windows.

[1689]