Add "Apply" button to "Go To Address". [1337]
I had to essentially reimplement the “Go To Address” functionality so
that the desired “Apply” button could work as expected. This was done
in the new GoToAddressWindow class.

Things should mostly work as of this commit, but I still need to get
the dialog to appear in the “right” place and handle errors better.