Make StateManager's processPropertyTable static.
Unfortunately, the changes made for Inquiry 1483 resulted in some
pretty bizarre bundle behavior (as seen in Inquiry 1768).

However, it turns out that there is no need for the stateManager
instance that was generating the NPE seen in 1483; simply making
StateManager’s “processPropertyTable” method static appears to fix
*both* inquiries.

I also had to back out of my local override of IDV’s StateManager due
to a strange interaction with the canonical (non-static) version of
processPropertyTable and my local override with its static
processPropertyTable. The More You Know!

[1768]