Enable control of value plot decluttering. [1740]
This adds the following methods to the _Layer wrapper object:

enableDeclutter()
disableDeclutter()
isDeclutterEnabled()
getDeclutterFilter()
setDeclutterFilter()

Currently the layer control GUI does not update if, for example,
decluttering is enabled and then disabled…but the actual display should
adjust as expected.

If you attempt to use any of the new methods with a layer other than
value plots, a NotImplementedError exception will be raised.