Bump JREs to 1.7.0_65. [1423]
Just grabbing the latest JREs from Oracle. It’s worth it just for the
increased speed with which the OSX installer unzips things! In fact,
the OSX installer is the only one that really changed; install4j seems
to bundle Java 7+ on OSX in a way that actually functions, so there’s
no need for our “osxjre.zip”.