Fix for new image chooser not appearing. [1801]
Take all of the choosers in "Under Development" as an example.

If all other choosers in "Under Development" are enabled, then McV
makes the assumption that the category is enabled in the preferences
and enables the new chooser.

If there is at least one disabled chooser in "Under Development", then
the category cannot be enabled in the preferences, and the new chooser
is disabled.