Don't auto-eval history menu items. [1799]
Prior to this commit, selecting an item from the Jython Shell history
menu would automatically evaluate the code. This seems like a bad user
experience thing in general, and it resulted in a new bug (they
automatically evaluated code wasn’t added to the history menu).

This commit changes the behavior of the history menu. Now when a
history menu item is selected, the results are added to the input
field. This gives the user a chance to make changes and also squashes
the bug described above!