Fix component group problems when capturing images
The code responsible for saving all displays for vector files didn’t
event know about component groups.

The raster file code knew about the component groups, but would attempt
to add *all* the ViewManagers in a IdvWindow to the list of “all
displays”.

Now both sections know of component groups, and will only consider the
ViewManagers that are within the displayed component holder.

[1837]