Add new IdvObjectStore wrapper class. [1730]
This is a first stab at a simple IdvObjectStore wrapper class that
allows McIDAS-V users to safely and easily read/write to main.xml.

I say "safely" because Jython performs implicit type conversions between
Java and Python types when possible. This is typically pretty convenient
when using Jython, but this behavior can easily call the wrong
(overloaded) IdvObjectStore get/put method...meaning that a call like
the following can end up writing a *long value* to the given property:

  idvObjStore.put("dummy.prop.name", Boolean.FALSE)

Note: the class name may not be final, as this is a nice starting point
for writing a "preferences" implementation entirely seperate from the
IDV.

Assorted niceties:
* All of the "get" methods *require* default values.
* Trying to use method names similar to java.util.prefs.Preferences.
* Has putObject/getObject methods that understand generics.