New image chooser class. [1825]
This new chooser is essentially just a copy of the existing
AddeImageParameterChooser. The key difference is that it creates an IDV
data source rather than McV's AddeImageParameterDataSource.

Also I'd like for it to eventually include the ability to wait until the
"absolute times" tab is visible before attempting contacting the server.

(i know there is an inquiry for that, but I can't recall the number)