Allow testCaptureImage to save KML files. [1660]
It's worth pointing out that KML has a significant limitation; all of
the generated images referenced within the KML must be located in the
same directory as the KML file itself. An alternative would be to use
absolute paths to the images, but that seems even worse than the current
usage of relative paths.

For now, testCaptureImage simply prints a warning message about the
limitation and suggests using KMZ instead.

It's kinda analogous to the IDV/McV zipped vs unzipped bundle situation
(my $0.02: unzipped should be deprecated!).