Add concept of "default" bundles to McV. [1756]
Default bundles are bundles that are shipped with the application. The
function largely the same as "toolbar" bundles, but *are not* managed
via the favorite bundles manager. I can provide the reasoning behind
this decision, but I'll spare everyone the tedium for now.

Creating default bundles is pretty easy; put your bundles in
edu/wisc/ssec/mcidasv/resources/defaultbundles/ and then edit
edu/wisc/ssec/mcidasv/resources/bundles.xml and reference the new
bundles (and be sure to set the "type" attribute to "default").

As of this commit there isn't a way to disable the new default bundles
from appearing, but I should have that fixed shortly.