Smarter ISL input check. [1906]
This removes the _isValidInput function, since for this particular use
case it is more straightforward to simply check "value is not None".
This works especially well since most parameters (currently) have
default values of None.

This also adds the checks (where appropriate) to all of the formatter
constructors.

Should've thought of this yesterday. :(