Disable display control menu item. [1891]
If the "enable adaptive resolution" preference is disabled, this commit
will disable the menu item.

I also opted to turn off the checkbox if adaptive resolution is
disabled, because otherwise users could see a grayed out menu item with
a checkbox. This implies (to me, at least) that adaptive resolution is
still on--but I'll defer to other people.