Store Jython history as arrays of bytes. [1831]
Old approach was to simply read/write the actual history string to
main.xml. Unfortunately, main.xml's encoding is ISO-8859-1, and this
leads to strange bugs when the history string contains non "latin"
characters.

New approach is to simply encode (UTF-8) the history string into an
array of bytes and write the byte array to main.xml.