Add Ant target for running arbitary script. [0]
This is something that I've been meaning to do for a LONG time now.
Unfortunately the syntax is a bit clunky, but here's how this works:

ant -Dpyfile="/path/to/script.py" jar.runscript

The jar.runscript target does essentially the same thing as
jar.runlarge.

Making a shell alias/function should be pretty straightforward.