Proof of concept workaround for blank images.
The idea is pretty simple:

1. Grab contents of display.

2. Try to find a non-blank pixel. If there is one, you're done.

3. Go to step 1.

This seems to work pretty well on my Linux machine (still need to try
Windows). Here's hoping it works for other people.

There are, of course, various optimizations and whatnot that can/will be
done--I just ran a few successful captureImage loops and wanted to get
this committed!

[1866]