Limit number of image capture retries. [1866]
If the initial image capture was unsuccessful, there will now be at most
10 retries. This avoids an infinite loop if the user has attempted to
capture a display that is actually blank (e.g. they've turned everything
off or zoomed in to a blank part of the display).

The retry limit can be bumped up if required, or maybe we can provide
users a way to set it.