Add "describe" method. [2039]
The idea is to make it easier to get a summary (of sorts) for a given
FlatField. See inquiry requests for how to use--I'm not yet convinced
that describe is useful enough to add to the default Jython imports.

The StatsTable change just makes the "fmtMe" method public and static.