Add a way to close windows via scripting. [2076]
I'd ideally like to be able to do something like:

  window = buildWindow(...)
  window.close()

But it's also really convenient that buildWindow returns a list of the
created _Displays. If the window was returned, we'd likely have to add
easy ways to access all of the window's displays (keeping in mind that
windows can contain multiple tabs, each with their own displays).