Use context managers to auto-close windows. [2077]
This adds a new Jython function "createWindow" (accepts same params as
buildWindow) that automatically cleans up a created window once users
finish using it.

It can be used like so:

  with createWindow() as panels:
      blah = loadADDEImage(**params1)
      print describe(blah)
      panels[0].createLayer('Image Display', blah)
      panels[0].captureImage('/tmp/hrm.png')

createWindow takes advantage of the "with statement" to create the
window via buildWindow, and then passes the results from buildWindow to
the code inside the "with block".

The window will be automatically closed once the code inside the "with
block" finishes executing (or if there was an exception).