Fix problem with notice dialog being too wide.
The explicit "check notice" action (this differs from the startup notice
check!) did not understand HTML prior to this commit, meaning that
longer notifications that relied on the "br" tag could appear in a
nighmarishly wide dialog box.

This also adds a method named "debugNotice" that is useful for testing
out a notification prior to sending it out. You can use it from the
Jython Shell like so:

getStaticMcv().getStateManager().debugNotice(noticeMessage, True, True)

[0]