Fix modification of JARs outside of repo. [2042]
The Ant targets ncidv.nolog and external.nolog would modify the JAR
files in "../IDV/lib" (relative to your mcv repo). This is especially
troublesome when "../IDV" happens to be a clone of the IDV repo.

This commit changes the two targets so that they only modify the copied
ncIdv.jar and external.jar within the "dist" subdirectory.