Add ability to hide "tab area" [2124]
This adds a new attribute to the "idv.component.group" skin tag (eww)
called "hideTabArea". The default value is "false", so no worries for
existing skins. If set to "true", the tab area should not be visible if
there is only a single tab in the tabbed pane.

There's still some tweaks that need to be done, but it's pretty close!