Fix for losing display list color after apply/ok.
The problem was that ViewManagers can override the control's color
setting. This change propagates the color change to all of the control's
ViewManagers, but not vice-versa.

In other words, setting the color in one ViewManager does not change the
control's colors (and thus the colors for the rest of the control's
ViewManagers).

[2068]