Add global palette checkbox to movie dialog. [798]
The checkbox will actually appear in both the capture movie "control"
dialog as well as the "save file" dialog.

Unfortunately, there does not appear to be a straightforward way of
enabling/disabling the checkboxes depending on the user selecting the
"Animated GIF" file mask (same thing for checking "Save As" textfield
for ".gif").