Create "About McIDAS-V" window more intelligently.
Profiling indicates that adding the contents of the "System Information"
tab to the text area is the reason why the "About" window could take a
little while to appear. It especially doesn't help that all of this was
happening on the EDT, which means that the GUI becomes unresponsive
until the window is visible!

This commit changes how the "System Information" tab is populated; when
the user makes selects the tab, a separate thread is fired off. This
thread gathers the system information, and when done, populates the tab
on the EDT.

This does result in a tiny delay while waiting for the tab to get
populated, but the GUI should feel more responsive overall.

[2242]