Handle missing directories better in captureImage.
captureImage gains an optional keyword, "createDirectories", that is a
boolean value (disabled by default). Some examples:

  activeDisplay().captureImage('/tmp/blah/test.png')

The above will raise a ValueError exception if any part of '/tmp/blah'
does not already exist.

  d = activeDisplay()
  d.captureImage('/tmp/blah/test.png', createDirectories=True)

This will attempt to create the directories needed to make
'/tmp/blah/test.png' a valid path.

This commit also improves the error message and some of the docs
associated with captureImage.

[2213]