Implement Bob's suggestions. [2337]
The "List All" widget now becomes a proper button, and I tried my hand
at improving the tooltip. I also turned the "Select Day" widget(s) into
a single button--but this required removing the icon to preserve
alignment. :(

I'd like to get the "N times selected" label and the bottom of the
timeline widget to align properly with the widgets on the right.
However, I'll defer to others on whether or not that should be a
priority.