Possible fix for ABI with unit=TEMP. [2226]
I'm tempted to wait to see if anyone notices the...additional feature...

But that would require writing a misleading commit message--and thus my
life as a programmer would be forfeit.

The preview window in the stable chooser/datasource will now attempt to
make use of related parameter aliases and defaults!