Add some trace logging to RGBCompControl. [2109]
I haven't had a lot of success replicating Bob's situation from
yesterday, but that's likely because things get crashy for me
with too many displays being created/destroyed...and I'm
reluctant to go in and start making "fixes" for problems
that may not exist.

With that in mind, this commit allows us to start gathering
at least some data about what may be happening--but it's
just counting iterations and time spent in initDone's RGB
while loops.