Add some tools to the McV monitor utility [2109]
When McV is running, the (*very basic*) interface available at
http://localhost:8788 offers a few new features:

* Attempt to show the Jython Shell.
* Attempt to show the Linear Combinations console.
* Set the global log level to TRACE.

The first two options are limited to showing their respective GUIs in
the active McV session, so security-wise it's no different from
clicking the "Jython Shell" menu item.

The difference between the Jython Shell and the Linear Combination
console is straightforward; the latter essentially does not try to do
any initialization (which is more useful than it may sound).