Fix problems with histogram and MultiSpecCntrl.
Creating a new MultiSpectralControl with a "bad" channel would cause
problems with the histogram--if it was even created at all!

This commit should fix that...as well as an issue where showing a valid
channel first, and then switching to a bad channel resulted in
everything *except* the actual chart getting updated to their
"NaN states".

This commit also changes the log level (WARN -> TRACE) for
ImagePlanViewControl's message about invalid histogram data. Previously
you'd get a somewhat alarming stack trace at the WARN level, even
though McV is operating as it should.

[702]