Potential fix for cursor/probe issues. [2416]
The default behavior when displaying new data is to "share" the time
across all layers (regardless of display). If the shared time is not
sufficiently close to one of the time steps in a given *layer*, users
will lose the ability to probe said layer.

The underlying problem is that if a layer has > 1 time step, the
animation code will always set the layer's "aniValue" to the shared
time, regardless of whether or not there was a match. This will cause
the probe to fail for the first time step, resulting in the layer
disallowing further probes.

That last bit is unfortunate, because if the "aniValue" will be changed
to a valid time step whenever the next time is displayed!