Create bare-bones mcidasv.rbi in user directory.
This fixes the problem where an idv.rbi would be created in the user
directory (even though McV would not look for it).

However, McV *does* look for mcidasv.rbi in the user directory. This
gives users the ability to add extra resources to McV in the event that
we have not yet added the ability to the McV GUI.

[2311]