Include CATransaction fix for macOS problems.
I'm not wild about having built an unofficial version of the macOS
JOGL "natives" JAR, but this is 100% worth it! Even more so because
it turns out that JOGL is surprisingly easy to build on macOS.
I'll add the instructions to the requests for this inquiry.

All credit for this fix belongs to Sven Gothel (maintainer of JOGL):

https://jogamp.org/bugzilla/show_bug.cgi?id=1299

Still...it would be nice to see this fix in the official JARs.

[2408]