Add Jython debugging class. [2109]
The strangest part of this problem is that we've seen users able to
"fix" their sessions by opening up the Jython Shell. This indicates that
something in the application depends on the state of *a* Jython
interpreter...

The intent behind LoudPyStringMap was to allow for certain variables to
be effectively immutable after initialization...but that ended up
causing activeDisplay().captureImage(...) to fail to "find" a
"islInterpreter"--but "print islInterpreter" immediately before and
after captureImage works just fine!

(this also caused the x-files theme to begin playing, but only once so
far. i'm chalking it up to threading problems.)