Write local grids dialog sizing work. [2211]
The most significant change is to GuiUtil's "showOkCancelDialog" method.

I added a parameter named "resizable" that controls whether or not the
resulting JDialog can be resized, and also made sure to set the
JDialog's preferred size to the preferred size of the contents
parameter.

GuiUtils also gained another "showOkCancelDialog" overload that ensures
all existing calls to showOkCancelDialog will behave the same as they
always have.

The actual dialog created in this case will be 1000x500, non-resizable,
and have the "Select Fields" as the title.