Fix some problems with histograms. [2470]
The main problem appears to be the IDV was using "McVHistogramWrapper",
while McV was using "McIDASVHistogramWrapper". The NPE seems to have
been a result of the "makeMenuItem" call to the McVHistogramWrapper
"saveImage" method--and that did not exist in McIDASVHistogramWrapper.

The only changes to the overrides are making the histoWrapper field in
ImagePlanViewControl protected, and changing several methods in
McVHistogramWrapper to be public rather than protected.