Make the file choosers more robust. [2237]
If the chooser's current directory is removed, the code should now be
smart enough to go to the nearest existing ancestor of the current
directory.

In other words, if you're looking at "/tmp/foo/bar/baz/lol" and
"/tmp/foo/bar" is deleted, the current directory will change to
"/tmp/foo".