Refactoring relative times. [2434]
Moved the most relevant changes out of TimesChooser and into
AddeChooser. This allows for the use of setState() and friends.

Put another way, the chooser should be better about handling invalid
input after this commit.

McVTextField can now (optionally) attempt to allow/deny the complete
text field contents, before the changes are applied. The previous
behavior was to allow/deny based on *only* the updated characters.