Make Jython Library handle filesystem events [1480]
The Jython Library (and shell) should now do a much better job of
reflecting the current state of the user's python directory.

So if a user were to add "blah.py" outside of McV, "blah.py" will be
added automatically and anything defined within the file will be
auto-imported into the Jython Shell.