Change ADDE preview logic. [2496]
This commit is rather messy, and I'm still not sure if the changes are
actually good...particularly hardcoding the line/element MAG to 1.

The good news is that the logic that actually speeds everything up is
pretty solid. The one remaining thing to do is to perform the ADDE
request off of the event dispatch thread, but this is proving to be a
bit more challenging.