Update installer JREs [2610]
Previous version was 1.8.0_112; new version is 1.8.0_152. 1.8.0_112 was
released over a year ago, so it's probably a good idea to update
anyway...but 1.8.0_152 has a new feature that allows McV to choose the
"unlimited strength" crypto policy, and this should enable testing to
move forward.

FWIW 1.8.0_162 is apparently going to default to the "unlimited" policy,
but it's still in testing.