Fix issue with way classpaths are created
Turns out that the "antcall" task creates a new scope for the target
task, meaning that the "Class-Path" attribute in McV's MANIFEST.MF
wasn't being built properly...which caused the user's guide to be
inaccessible.

This commit adds the "ant-contrib" project to the build dependencies
because it includes the "antcallback" task (which allows us to get
properties out of the scope).

[2661]