Remove platform-specific JARs. [2661]
Owing to the fact that we now rely so heavily on creating the classpath
at *build time*, we were inadvertently creating platform-specific sets
of JAR files! Mea culpa! :(

The fix is to use the JogAmp JAR file that contains all of the various
JOGL dependencies and simply remove all of the platform-specific JARs
from the McV repo.

This also cleans up build.xml a bit.