Don't show toolbar actions that use monkey icon
The "monkey" icon doubles as our "no icon has been associated with this
action" icon, and generally the actions using that icon are somewhat
buggy.

For the record, if the user had previously decided to add one of these
actions to their toolbar, they will still be able to use that toolbar
button.

[1220]