Change to anomalyFromTimeMeans function [2732]
There were two functions called "anomalyFromTimeMeans"  One accepted one parameter, and the other accepted two.  The name of the function that accepts two parameters was changed to "anomalyFromTimeMeansType".  This is done to get formulas that need the function with two parameters working.

Additionally, a typo was fixed where "montly" was changed to "monthly".  This was done to get the "monthly" formula working.

I also added an "if block" so the meanType of "yearly" can be used.  "yearly" is the default of one of the UI formulas.