Setting isenduser to false for two formulas [2732]
The formulas "Change Unit of a parameter" and "Compute anomaly from time means at specified time type" were fully commented out since I can't get them working, but the result of this was to bring them back under the MapesCollection tree.  To avoid this, I'm not commenting out the formulas, but setting isenduser to false so they don't list in the Field Selector.