Add satellites and alphabetize list [2746]
This file is used to obtain the swath width of different satellites for the Satellite Orbit Track chooser.
Several satellites from CelesTrak and noaaport.ssec.wisc.edu/POESNAV were added to the list.  Some swath widths were added to satellites that were listed but didn't include a width.  The listing is now in alphabetical order based on satellite name.