Fix problem with redundant ground stations.
When trying to save off the currently plotted ground stations, we use a
list ("stations") to ease the process. The "getStations()" method simply
iterates through the plotted stations, and adds them to the "stations"
list.

The problem described in Request 18, Bug 4 is that we were saving
multiple bundles from the same control--without clearing the stations
list! This meant that every call to "getStations()" would result in all
of the plotted stations being added to the existing contents of the
station list.

The fix was pretty simple: clear the "stations" list at the start of
"getStations()".

[2746]