Added more variables that weren't aliased [2791]
Added a variety of variables from HRRR, SREF, NDFD (NOAAPORT and CONDUIT), RTMA, and National Blend models to this file so they are correctly aliased.  These data sources are all available through the default catalog in the Gridded>Remote chooser.  I updated the Summary of the inquiry to reflect that this inquiry now addresses more than just GFS data.