Adding GOES 16/17 datasets to lead server [2755]
The GOES 16 datasets had been previously removed from the lead.unidata.ucar.edu list since it was mentioned that this server should only be used as a last resort.  Tom Y just suggested that the GOES 16/17 datasets should be added back to McV.

Also of note is that Tom Y plans on getting the GOES 16/17 datasets working this summer on adde.ssec.wisc.edu.  They need a newer machine with more memory to do this, but it's high on their priority list.  More specifics on this in the Request of this inquiry.