Estimate time spent init-ing Jython interpreters.
Just using System.nanoTime() as a way to get quick, *ball park*
estimates of how much startup time is spent initializing Jython
interpreters.

The accurate way to do this would be to use JMH, but the idea behind
this estimate is to provide a way for MUG to determine if certain
aspects of McV's startup are remaining steady, slowing down, or speeding
up. So we'd prefer running at every startup and just watching the
trends, and JMH isn't as suited to that.

[2751]