Fix problem with not using correct local ADDE port
The root of the problem is that the list of available ADDE entries was
being created once, which meant that the if the local ADDE port changed
sometime during the session we'd be out of luck.

The change is to no longer supply the port number in EntryTransforms,
and for AddeServer's overridden "getName()" method to query the server
manager for the port to use--but only if the AddeServer's "isLocal"
field is set to true.

[2821]