Fix problem with leading '/' on Windows [2721]
Probably worth repeating that this is a Windows-specific fix.

It looks like the problem is downwind of the fact that the NCML
"%location%" macro uses URLs, meaning that the following path:

Z:\Users\bob\test.txt

becomes the following in the MiRS plugin's NCML location:

file:/Z:/Users/bob/test.txt

By the time we get to TropomiIOSP's isValidFile(...) method, that path
has turned into the following:

/Z:/Users/bob/test.txt

All the fix does is check to see if we're on Windows and if we match a
simple regex. If so, just remove the first character from the path being
checked.