Spelling out calibration for ABI formulas [2836]
Changed from saying the first character of the calibration (e.g. T) to the full name of the calibration (e.g. Temperature).  This is for consistency with how the VIIRS formulas will need to be done since sometimes they require Reflectance and sometimes they require Radiance, so simply using "R" in the label wouldn't work.  This label is seen in the Field Selector window when manually selecting the bands/calibrations for an RGB or channel difference formula.