Potential fix for weird fields panel bug [2860]
I'm not sure that this fixes 2860, as I couldn't reliably replicate the
problem in the first place...but I noticed that the
DataSelector#addDataSource(DataSource, boolean) calls were not
happening on the EDT (and they definitely should be).

At any rate, I've yet to get the problem to happen again with this
commit.