Resource for plotting more city labels [2864]
Adds 15,493 new city labels scattered around the world.  Big improvement in city label density especially outside of the CONUS.  The city listing was originally obtained from https://simplemaps.com/data/world-cities and parsed with a Python script to get it in this format.  This process is outlined in the Request of the inquiry.