Source of city locations requires license [2864]
The city database is licensed under the Creative Commons Attribution 4.0 license, which is described and linked to in this license file.  This license was included in the download of the city CSV file from Simplemaps.com.

Note of this will also be added to the License & Copyright page of the User's Guide.