Potential fix for intermittent bundle NPE [2861]
Extraneous DataOperands were being created (named like "D1", "D2", etc)
which lacked corresponding DataChoices to associate with, causing an NPE

I've tested this with a few formulas and Bob's example bundle, but there
are a lot of moving parts in this code and so there's a chance I may
have made incorrect assumptions. Put another a way: this is tentative!