Fix problem with install4j configuration [2839]
When dealing with multiple platforms, install4j wants you to create sets
of files (e.g. Set A for Windows 64, Set B for Java 11 Linux, etc). The
platform-specific installers ultimately use these file sets to determine
the files that get installed.

The problem in this case is that both the non-JPMS and JPMS versions of
runMcV were being installed, with the latter overwriting the former.

The fix is easy; just exclude the macOS JPMS file set from the non-JPMS
installer (and vice versa).